refactor(web-ui): decompose Universes god-component
Split the 929-line Universes page into a slim stateful container plus
presentational pieces under components/domain/universes/:
sidebar, config/constituents/discovery/screener tab panels, and a
shared constants/helpers module. Container keeps all state, queries,
and mutations; rendered DOM and public interface unchanged.
